+++ axum-extra/CHANGELOG.md
@@ -7,9 +7,11 @@ and this project adheres to [Semantic Versioning].
 
 # Unreleased
 
+- **added:** Implement `Deref` and `DerefMut` for built-in extractors ([#1922])
 - **added:** Add `OptionalPath` extractor ([#1889])
 
 [#1889]: https://github.com/tokio-rs/axum/pull/1889
+[#1922]: https://github.com/tokio-rs/axum/pull/1922
 
 # 0.7.2 (22. March, 2023)